refactor: split data/fs/report_parser/runs.py into focused modules
